Move useWhatsabi to its own file
